app/main.py
import os
import torch
from PIL import Image
from typing import Dict, Any

class FashionPipeline:
    def __init__(self, config: Dict[str, Any] = None):
        """
        Initialize the Fashion Pipeline with configuration
        Args:
            config: Configuration dictionary for models and parameters
        """
        self.config = config or {}
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self._initialize_models()

    def _initialize_models(self):
        """Initialize required AI models"""
        # TODO: Initialize SAM2 and Stable Diffusion models
        pass

    def process_image(self, image_path: str) -> Image.Image:
        """
        Process a clothing image through the full pipeline
        Args:
            image_path: Path to the input clothing image
        Returns:
            PIL.Image: Generated fashion model wearing the input clothing
        """
        # Load and preprocess input image
        input_image = self._load_image(image_path)
        
        # Run SAM2 segmentation
        masks = self._segment_clothing(input_image)
        
        # Refine segmentation mask
        refined_mask = self._refine_mask(masks)
        
        # Extract clothing features
        features = self._extract_features(input_image, refined_mask)
        
        # Generate fashion model
        result = self._generate_fashion_model(input_image, refined_mask, features)
        
        return result

    def _load_image(self, image_path: str) -> Image.Image:
        """Load and preprocess input image"""
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"Image not found: {image_path}")
        
        image = Image.open(image_path)
        # TODO: Add preprocessing steps
        return image

    def _segment_clothing(self, image: Image.Image):
        """Perform SAM2 segmentation on input image"""
        # TODO: Implement SAM2 segmentation
        pass

    def _refine_mask(self, masks):
        """Refine the segmentation mask"""
        # TODO: Implement mask refinement
        pass

    def _extract_features(self, image: Image.Image, mask):
        """Extract features from the segmented clothing"""
        # TODO: Implement feature extraction
        pass

    def _generate_fashion_model(self, image: Image.Image, mask, features):
        """Generate fashion model using Stable Diffusion"""
        # TODO: Implement generation pipeline
        pass

def main():
    # Example usage
    pipeline = FashionPipeline()
    result = pipeline.process_image("path_to_clothing_image.jpg")
    result.save("output.jpg")

if __name__ == "__main__":
    main()