requirements.txt
torch>=2.0.0
segment-anything
diffusers>=0.19.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
accelerate
safetensors
flask>=2.0.0
werkzeug>=2.0.0